View refactoring
